Update _.last
The function accepts an array and an optional n value, which can be
used to indicate how many elements from the end of the array you would
like to receive. If n does not exist, the last element of the array
will be returned. If n does exist, an n character array will be returned
that concludes with the final element. If n is greater than the length
of the array, the entire array will be returned.